requirements.txt
opencv-python
numpy
ultralytics
google-generativeai
python-dotenv